promoted optimise to abstract nest and added docs
